chore(json-ld): move util methods to AbstractJsonLdTransformer
