+++ extensions/common/json-ld/src/main/java/org/eclipse/edc/jsonld/transformer/to/JsonObjectToCatalogTransformer.java
@@ -25,12 +25,8 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import static java.lang.String.format;
-import static org.eclipse.edc.jsonld.transformer.JsonLdFunctions.nodeId;
-import static org.eclipse.edc.jsonld.transformer.JsonLdFunctions.nodeType;
 import static org.eclipse.edc.jsonld.transformer.JsonLdNavigator.visitProperties;
 import static org.eclipse.edc.jsonld.transformer.Namespaces.DCAT_SCHEMA;
-import static org.eclipse.edc.jsonld.transformer.TransformerUtil.transformArrayOrObject;
-import static org.eclipse.edc.jsonld.transformer.TransformerUtil.transformGenericProperty;
 
 public class JsonObjectToCatalogTransformer extends AbstractJsonLdTransformer<JsonObject, Catalog> {
