+++ extensions/common/json-ld/src/main/java/org/eclipse/edc/jsonld/transformer/to/JsonObjectToPermissionTransformer.java
@@ -27,8 +27,6 @@ import org.jetbrains.annotations.Nullable;
 
 import static org.eclipse.edc.jsonld.transformer.JsonLdNavigator.visitProperties;
 import static org.eclipse.edc.jsonld.transformer.Namespaces.ODRL_SCHEMA;
-import static org.eclipse.edc.jsonld.transformer.TransformerUtil.transformArrayOrObject;
-import static org.eclipse.edc.jsonld.transformer.TransformerUtil.transformObject;
 
 public class JsonObjectToPermissionTransformer extends AbstractJsonLdTransformer<JsonObject, Permission> {
